fix annoying warning
Somehow the plugin generates some warning (possibly only when combined with babel typescript plugin) that can be annoying during dev.
```
The exported identifier "_$Component" is not declared in Babel's scope tracker
as a JavaScript value binding, and "@babel/plugin-transform-typescript"
never encountered it as a TypeScript type declaration.
It will be treated as a JavaScript value.

This problem is likely caused by another plugin injecting
"_$Component" without registering it in the scope tracker. If you are the author
 of that plugin, please use "scope.registerDeclaration(declarationPath)".
```
This should fix the issue.